+++ src/test/java/com/fhsh/daitda/UserApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fhsh.user;
+package com.fhsh.daitda;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
